Admin klar

+++ router/admin/admin.js
@@ -5,28 +5,32 @@ const verifyToken = require("../verifyToken")
 const router = express.Router();
 
 router.route("/admin")
-    .get(verifyToken , async (req, res) => {
-    if(req.user.admin == false) return res.redirect("/")
+    .get(verifyToken, async (req, res) => {
+        const user = await User.findOne({ _id: req.user.user._id })
 
-        const sortName = req.query.name;
-        const sortPrice = req.query.price;
-        const sortCategory = req.query.category;
-        const queryExist = req.query.page;
+        if (req.user.user.admin == true) {
+            const sortName = req.query.name;
+            const sortPrice = req.query.price;
+            const sortCategory = req.query.category;
+            const queryExist = req.query.page;
 
-        let productQuantity = await Candy.find().countDocuments();
+            let productQuantity = await Candy.find().countDocuments();
 
-        const page = +req.query.page || 1;
-        const productsPerPage = 4;
-        let pageQuantity = await Candy.find().countDocuments() / productsPerPage;
-        pageQuantity = Math.ceil(pageQuantity);
+            const page = +req.query.page || 1;
+            const productsPerPage = 4;
+            let pageQuantity = await Candy.find().countDocuments() / productsPerPage;
+            pageQuantity = Math.ceil(pageQuantity);
 
-        const pageCount = Math.ceil(productQuantity / productsPerPage)
+            const pageCount = Math.ceil(productQuantity / productsPerPage)
 
-        const findCandy = await Candy.find().collation({ locale: "sv", strength: 2 }).sort({ name: sortName }).skip(productsPerPage * (page - 1)).limit(productsPerPage);
-        
-        res.render("admin/adminProduct", { findCandy, page, pageQuantity, productsPerPage, queryExist, pageCount, title: "Admin - Lasses Lakrits" })
+            const findCandy = await Candy.find().collation({ locale: "sv", strength: 2 }).sort({ name: sortName }).skip(productsPerPage * (page - 1)).limit(productsPerPage);
+
+            res.render("admin/adminProduct", { user, findCandy, page, pageQuantity, productsPerPage, queryExist, pageCount, title: "Admin - Lasses Lakrits" })
+        } else {
+            res.send("Du har inte rättigheter för den här sidan");
+        }
+
     })
-
     .post(async (req, res) => {
         await new Candy({
             name: req.body.name,
@@ -35,7 +39,10 @@ router.route("/admin")
             category: req.body.category,
             color: req.body.color,
             createdByAdmin: req.body.createdByAdmin,
-            img: req.body.img
+            img: req.body.img,
+            // user: "5e763d50c5002e07b42894ab"
+            user: req.body.user
+
 
         }).save((error, success) => {
             if (error) {
@@ -48,16 +55,28 @@ router.route("/admin")
     });
 
 router.route("/delete/:id")
-    .get(async (req, res) => {
-        await Candy.deleteOne({ _id: req.params.id });
-        res.redirect("/admin");
+    .get(verifyToken, async (req, res) => {
+        const user = await User.findOne({ _id: req.user.user._id });
+
+        if (req.user.user.admin == true) {
+            await Candy.deleteOne({ _id: req.params.id });
+            res.redirect("/admin");
+        } else {
+            res.send("Du har inte rättigheter för den här sidan");
+        }
     })
 
 
 router.route("/update/:id")
-    .get(async (req, res) => {
-        const updateCandy = await Candy.findById({ _id: req.params.id });
-        res.render("admin/updateProduct", { updateCandy, title: "Update Candy" });
+    .get(verifyToken, async (req, res) => {
+        const user = await User.findOne({ _id: req.user.user._id });
+
+        if (req.user.user.admin == true) {
+            const updateCandy = await Candy.findById({ _id: req.params.id });
+            res.render("admin/updateProduct", { updateCandy, title: "Update Candy" });
+        } else {
+            res.send("Du har inte rättigheter för den här sidan");
+        }
     })
     .post(async (req, res) => {
         await Candy.updateOne({ _id: req.params.id },
@@ -77,35 +96,47 @@ router.route("/update/:id")
 
 
 router.route("/admin2")
-    .get(async (req, res) => {
-        const sortName = req.query.name;
-        const sortAdmin = req.query.admin;
+    .get(verifyToken, async (req, res) => {
+        const user = await User.findOne({ _id: req.user.user._id });
 
-        const queryExist = req.query.page;
+        if (req.user.user.admin == true) {
+            const sortName = req.query.name;
+            const sortAdmin = req.query.admin;
 
-        let userQuantity = await User.find().countDocuments();
+            const queryExist = req.query.page;
 
-        const page = +req.query.page || 1;
-        const usersPerPage = 4;
-        let pageQuantity = await User.find().countDocuments() / usersPerPage;
-        pageQuantity = Math.ceil(pageQuantity);
+            let userQuantity = await User.find().countDocuments();
 
-        const pageCount = Math.ceil(userQuantity / usersPerPage)
+            const page = +req.query.page || 1;
+            const usersPerPage = 4;
+            let pageQuantity = await User.find().countDocuments() / usersPerPage;
+            pageQuantity = Math.ceil(pageQuantity);
 
-        let onlyUsers = { admin: false };
-        const findUsers = await User.find(onlyUsers).collation({ locale: "sv", strength: 2 }).sort({ name: sortName }).skip(usersPerPage * (page - 1)).limit(usersPerPage);
+            const pageCount = Math.ceil(userQuantity / usersPerPage)
 
-        let onlyAdmins = { admin: true };
-        const findAdmins = await User.find(onlyAdmins).collation({ locale: "sv", strength: 2 }).sort({ admin: sortAdmin }).skip(usersPerPage * (page - 1)).limit(usersPerPage);
+            let onlyUsers = { admin: false };
+            const findUsers = await User.find(onlyUsers).collation({ locale: "sv", strength: 2 }).sort({ name: sortName }).skip(usersPerPage * (page - 1)).limit(usersPerPage);
 
-        res.render("admin/adminUser", { findUsers, findAdmins, page, pageQuantity, usersPerPage, queryExist, pageCount, title: "Admin - Lasses Lakrits" })
+            let onlyAdmins = { admin: true };
+            const findAdmins = await User.find(onlyAdmins).collation({ locale: "sv", strength: 2 }).sort({ admin: sortAdmin }).skip(usersPerPage * (page - 1)).limit(usersPerPage);
+
+            res.render("admin/adminUser", { user, findUsers, findAdmins, page, pageQuantity, usersPerPage, queryExist, pageCount, title: "Admin - Lasses Lakrits" })
+        } else {
+            res.send("Du har inte rättigheter för den här sidan");
+        }
     })
 
 
 router.route("/updateUser/:id")
-    .get(async (req, res) => {
-        const findUser = await User.findById({ _id: req.params.id });
-        res.render("admin/updateUser", { findUser, title: "Update Admin" });
+    .get(verifyToken, async (req, res) => {
+        const user = await User.findOne({ _id: req.user.user._id });
+
+        if (req.user.user.admin == true) {
+            const findUser = await User.findById({ _id: req.params.id });
+            res.render("admin/updateUser", { findUser, title: "Update Admin" });
+        } else {
+            res.send("Du har inte rättigheter för den här sidan");
+        }
     })
     .post(async (req, res) => {
         await User.updateOne({ _id: req.params.id },
@@ -118,6 +149,6 @@ router.route("/updateUser/:id")
             }, { runValidators: true });
         res.redirect("/admin2");
     })
-    
+
 
 module.exports = router;
